parse request params of multiply controller to float

+++ day-02/api/controller/computation.js
@@ -1,8 +1,8 @@
 "use strict";
 
 const multiplyTwoNumbers = (req, res) => {
-  const firstNum = parseInt(req.params.firstNum);
-  const secondNum = parseInt(req.query.secondNum);
+  const firstNum = parseFloat(req.params.firstNum);
+  const secondNum = parseFloat(req.query.secondNum);
 
   res.status(200).json({
     firstNum,
